fix compile warning
git-svn-id: http://192.168.20.11/sequoiadb/trunk@41281 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/cls/clsShardSession.cpp
@@ -1790,7 +1790,7 @@ namespace engine
       {
          MsgOpTransBegin *pTransBegin = ( MsgOpTransBegin* )msg ;
          /// Old trans begin msg is only a MsgHeader
-         if ( msg->messageLength > sizeof( MsgHeader ) &&
+         if ( msg->messageLength > (INT32)sizeof( MsgHeader ) &&
               DPS_INVALID_TRANS_ID != pTransBegin->transID &&
               0 != DPS_TRANS_GET_NODEID( pTransBegin->transID ) )
          {
